rename log group event source construct file
